+++ src/main/java/framework/commonfunctions/CommonFunctions.java
@@ -347,11 +347,32 @@ public class CommonFunctions {
 	 * @param maxTimeout the max timeout in seconds
 	 * @throws Exception the exception
 	 */
-	public void waitForInvisibilityOfElement(WebDriver driver, WebElement element, int maxTimeout) throws Exception {
+	public boolean waitForInvisibilityOfElement(WebDriver driver, WebElement element, int maxTimeout) throws Exception {
+		return waitForInvisibilityOfElement(driver, element, CommonVariables.MIN_TIMEOUT, maxTimeout);
+	}
+
+	/**
+	 * Wait for invisibility of element. Method will wait for
+	 * {@link CommonVariables#MIN_TIMEOUT} before checking for element invisibility.
+	 *
+	 * @param driver     the {@link org.openqa.selenium.WebDriver WebDriver}
+	 * @param element    the {@link org.openqa.selenium.WebElement element}
+	 * @param initialWaitTime the time to be waited for the element to be present in
+	 * 	 *                        seconds
+	 * @param maxTimeout the max timeout in seconds
+	 * @throws Exception the exception
+	 * @return the status of element invisibility<br>
+	 *	<b>Note:</b><br>
+	 *	   <ul>
+	 *	       <li>true: if the element is invisible</li>
+	 *	       <li>false: if the element is visible</li>
+	 *	   </ul>
+	 */
+	public boolean waitForInvisibilityOfElement(WebDriver driver, WebElement element, int initialWaitTime, int maxTimeout) throws Exception {
 		this.logAccess.getLogger().info("waiting for element to be invisible  :- " + element);
 
 		try {
-			waitForElement(driver, element, ExpectedConditionsEnums.VISIBLE, CommonVariables.MIN_TIMEOUT);
+			waitForElement(driver, element, ExpectedConditionsEnums.VISIBLE, initialWaitTime);
 		} catch (Exception ignoreException) {
 			// ignore the exception and continue with the script
 		}
@@ -360,19 +381,19 @@ public class CommonFunctions {
 		int waitingSeconds = maxTimeout * 1000;
 		long endTimestamp = currentTimestamp + waitingSeconds;
 
-		boolean isElementInvisible = true;
+		boolean isElementVisible = true;
 
 		this.logAccess.getLogger().info("End timestamp for Invisibility of an Element is " + endTimestamp);
 
 		try {
 
-			while ((new Date()).getTime() < endTimestamp && isElementInvisible) {
+			while ((new Date()).getTime() < endTimestamp && isElementVisible) {
 
-				isElementInvisible = isElementPresent(driver, element, CommonVariables.NO_TIMEOUT);
+				isElementVisible = isElementPresent(driver, element, CommonVariables.NO_TIMEOUT);
 
-				if (isElementInvisible) {
+				if (isElementVisible) {
 					// Checking if element is visible though it is in the DOM.
-					isElementInvisible = element.isDisplayed();
+					isElementVisible = element.isDisplayed();
 
 				}
 				Thread.sleep(500);
@@ -381,8 +402,9 @@ public class CommonFunctions {
 			// intentionally left it blank (we can ignore the above exceptions when waiting
 			// for element in-visibility)
 		}
-
+		return !isElementVisible;
 	}
+
 
 	/**
 	 * Wait for invisibility of element by locator. Method will wait for
@@ -390,41 +412,12 @@ public class CommonFunctions {
 	 *
 	 * @param driver     the {@link org.openqa.selenium.WebDriver WebDriver}
 	 * @param byLocator  the by locator
-	 * @param maxTimeout the max timeout in seconds
+	 * @param maxTimeout the time to be waited for the element to be present in
+	 * 	 *                        seconds
 	 * @throws Exception the exception
 	 */
-	public void waitForInvisibilityOfElement(WebDriver driver, By byLocator, int maxTimeout) throws Exception {
-		this.logAccess.getLogger().info("waiting for element to be invisible  :- " + byLocator);
-
-		try {
-			waitForElement(driver, byLocator, ExpectedConditionsEnums.PRESENCE, CommonVariables.MIN_TIMEOUT);
-		} catch (Exception ignoreException) {
-			// ignore the exception and continue with the script
-		}
-
-		long currentTimestamp = (new Date()).getTime();
-		int waitingSeconds = maxTimeout * 1000;
-		long endTimestamp = currentTimestamp + waitingSeconds;
-
-		boolean isElementInvisible = true;
-
-		this.logAccess.getLogger().info("End timestamp for Invisibility of an Element is " + endTimestamp);
-		try {
-			while ((new Date()).getTime() < endTimestamp && isElementInvisible) {
-
-				isElementInvisible = isElementPresent(driver, byLocator, CommonVariables.NO_TIMEOUT);
-
-				if (isElementInvisible) {
-					// Checking if element is visible though it is in the DOM.
-					isElementInvisible = isElementDisplayed(driver, byLocator, CommonVariables.NO_TIMEOUT);
-				}
-				Thread.sleep(500);
-			}
-		} catch (NoSuchElementException | StaleElementReferenceException ignoreException) {
-			// intentionally left it blank (we can ignore the above exceptions when waiting
-			// for element in-visibility)
-		}
-
+	public boolean waitForInvisibilityOfElement(WebDriver driver, By byLocator, int maxTimeout) throws Exception {
+		return waitForInvisibilityOfElement(driver,byLocator, CommonVariables.MIN_TIMEOUT, maxTimeout);
 	}
 
 	/**
@@ -438,12 +431,12 @@ public class CommonFunctions {
 	 * @param maxTimeout      the max timeout in seconds
 	 * @throws Exception the exception
 	 */
-	public void waitForInvisibilityOfElement(WebDriver driver, By byLocator, int initialWaitTime, int maxTimeout)
+	public boolean waitForInvisibilityOfElement(WebDriver driver, By byLocator, int initialWaitTime, int maxTimeout)
 			throws Exception {
 		this.logAccess.getLogger().info("waiting for element to be invisible  :- " + byLocator);
 
 		try {
-			waitForElement(driver, byLocator, ExpectedConditionsEnums.PRESENCE, initialWaitTime);
+			waitForElement(driver, byLocator, ExpectedConditionsEnums.VISIBLE, initialWaitTime);
 		} catch (Exception ignoreException) {
 			// ignore the exception and continue with the script
 		}
@@ -452,17 +445,17 @@ public class CommonFunctions {
 		int waitingSeconds = maxTimeout * 1000;
 		long endTimestamp = currentTimestamp + waitingSeconds;
 
-		boolean isElementInvisible = true;
+		boolean isElementVisible = true;
 
 		this.logAccess.getLogger().info("End timestamp for Invisibility of an Element is " + endTimestamp);
 		try {
-			while ((new Date()).getTime() < endTimestamp && isElementInvisible) {
+			while ((new Date()).getTime() < endTimestamp && isElementVisible) {
 
-				isElementInvisible = isElementPresent(driver, byLocator, CommonVariables.NO_TIMEOUT);
+				isElementVisible = isElementPresent(driver, byLocator, CommonVariables.NO_TIMEOUT);
 
-				if (isElementInvisible) {
+				if (isElementVisible) {
 					// Checking if element is visible though it is in the DOM.
-					isElementInvisible = isElementDisplayed(driver, byLocator, CommonVariables.NO_TIMEOUT);
+					isElementVisible = isElementDisplayed(driver, byLocator, CommonVariables.NO_TIMEOUT);
 				}
 				Thread.sleep(500);
 			}
@@ -470,7 +463,7 @@ public class CommonFunctions {
 			// intentionally left it blank (we can ignore the above exceptions when waiting
 			// for element in-visibility)
 		}
-
+		return !isElementVisible;
 	}
 
 	/**
@@ -485,9 +478,10 @@ public class CommonFunctions {
 		this.logAccess.getLogger().info("checking if element is present  :- " + element);
 		// wait for the {@link org.openqa.selenium.WebElement element} to present
 		try {
-			waitForElement(driver, element, ExpectedConditionsEnums.PRESENCE, CommonVariables.MED_TIMEOUT);
+			WebElement targetEle = waitForElement(driver, element, ExpectedConditionsEnums.PRESENCE, CommonVariables.MED_TIMEOUT);
 			// return true if the element is present
-			return true;
+			return targetEle != null;
+
 		} catch (Exception e) {
 			// return false if the element is not
 			// present
@@ -508,9 +502,11 @@ public class CommonFunctions {
 		// wait for the {@link org.openqa.selenium.WebElement element} to present for
 		// specified time
 		try {
-			waitForElement(driver, element, ExpectedConditionsEnums.PRESENCE, maxTimeout);
+			WebElement targetEle = waitForElement(driver, element, ExpectedConditionsEnums.PRESENCE, maxTimeout);
+
 			// return true if the element is present
-			return true;
+			return targetEle != null;
+
 		} catch (Exception e) {
 			// return false if the element is not
 			// present
@@ -530,9 +526,11 @@ public class CommonFunctions {
 		// wait for the {@link org.openqa.selenium.WebElement element} to present for
 		// specified time
 		try {
-			waitForElement(driver, byLocator, ExpectedConditionsEnums.PRESENCE, CommonVariables.MED_TIMEOUT);
+			WebElement targetEle = waitForElement(driver, byLocator, ExpectedConditionsEnums.PRESENCE, CommonVariables.MED_TIMEOUT);
+
 			// return true if the /element is visible
-			return true;
+			return targetEle != null;
+
 		} catch (Exception e) {
 			// return false if the element is not
 			// visible
@@ -553,9 +551,9 @@ public class CommonFunctions {
 		// wait for the {@link org.openqa.selenium.WebElement element} to present for
 		// specified time
 		try {
-			waitForElement(driver, byLocator, ExpectedConditionsEnums.PRESENCE, maxTimeout);
+			WebElement targetEle = waitForElement(driver, byLocator, ExpectedConditionsEnums.PRESENCE, maxTimeout);
 			// return true if the /element is visible
-			return true;
+			return targetEle != null;
 		} catch (Exception e) {
 			// return false if the element is not
 			// visible
@@ -576,9 +574,11 @@ public class CommonFunctions {
 		// wait for the {@link org.openqa.selenium.WebElement element} to present for
 		// specified time
 		try {
-			waitForElement(driver, byLocator, ExpectedConditionsEnums.VISIBLE, maxTimeout);
+			WebElement targetEle = waitForElement(driver, byLocator, ExpectedConditionsEnums.VISIBLE, maxTimeout);
+
 			// return true if the element is visible
-			return true;
+			return targetEle != null;
+
 		} catch (Exception e) {
 			// return false if the element is not
 			// visible
@@ -597,9 +597,11 @@ public class CommonFunctions {
 		this.logAccess.getLogger().info("checking if element is displayed  :- " + byLocator);
 		// wait for the element to present for specified time
 		try {
-			waitForElement(driver, byLocator, ExpectedConditionsEnums.VISIBLE, CommonVariables.MED_TIMEOUT);
+			WebElement targetEle = waitForElement(driver, byLocator, ExpectedConditionsEnums.VISIBLE, CommonVariables.MED_TIMEOUT);
+
 			// return true if the element is visible
-			return true;
+			return targetEle != null;
+
 		} catch (Exception e) {
 			// return false if the element is not
 			// visible
@@ -618,9 +620,9 @@ public class CommonFunctions {
 	public boolean isElementDisplayed(WebDriver driver, WebElement element, int maxTimeout) {
 		this.logAccess.getLogger().info("checking if element is displayed  :- " + element);
 		try {
-			waitForElement(driver, element, ExpectedConditionsEnums.VISIBLE, maxTimeout);
+			WebElement targetEle = waitForElement(driver, element, ExpectedConditionsEnums.VISIBLE, maxTimeout);
 			// return true if the element is visible
-			return true;
+			return targetEle != null;
 		} catch (Exception e) {
 			// return false if the element is not
 			// visible
@@ -638,9 +640,10 @@ public class CommonFunctions {
 	public boolean isElementDisplayed(WebDriver driver, WebElement element) {
 		this.logAccess.getLogger().info("checking if element is displayed  :- " + element);
 		try {
-			waitForElement(driver, element, ExpectedConditionsEnums.VISIBLE, CommonVariables.MED_TIMEOUT);
+			WebElement targetEle = waitForElement(driver, element, ExpectedConditionsEnums.VISIBLE, CommonVariables.MED_TIMEOUT);
 			// return true if the element is visible
-			return true;
+			return targetEle != null;
+
 		} catch (Exception e) {
 			// return false if the element is not
 			// visible
@@ -1017,7 +1020,11 @@ public class CommonFunctions {
 		this.logAccess.getLogger().info("Element :- " + element);
 		// get the element
 		WebElement tempElement = getElement(driver, element, maxTimeOut);
-		if (tempElement.getAttribute("type").equals("password")) {
+
+		// Check if the tag name is input, there are cases where we can enter the
+		// data but the field is not input and it's not possible to check the type
+		// attribute in those scenarios.
+		if (tempElement.getAttribute("tagName").equalsIgnoreCase("input") && tempElement.getAttribute("type").equals("password")) {
 			this.logAccess.getLogger().info("Masked value :- " + "x".repeat(value.length()));
 		} else {
 			this.logAccess.getLogger().info("value :- " + value);
@@ -1042,27 +1049,6 @@ public class CommonFunctions {
 		}
 		// un-highlight
 		unHighlightElement(driver, tempElement, originalStyle);
-	}
-
-	/**
-	 * Input value.
-	 *
-	 * @param driver              the {@link org.openqa.selenium.WebDriver
-	 *                            WebDriver}
-	 * @param element             the {@link org.openqa.selenium.WebElement element}
-	 * @param value               the value to be set in the element
-	 * @param isCaptureScreenshot toggle to capture screenshot
-	 * @param screenShotName      the screen shot name <br>
-	 *                            Date time Stamp will be <i>prepended</i> to the
-	 *                            screenshot name by default.<br>
-	 *                            Note: Use {@link #screenShotsPath screenShotsPath}
-	 *                            setter to set the path where you want to store the
-	 *                            screenshots.
-	 * @throws Exception the exception
-	 */
-	public void inputValue(WebDriver driver, WebElement element, String value, boolean isCaptureScreenshot,
-						   String screenShotName) throws Exception {
-		inputValue(driver, element, value, isCaptureScreenshot, screenShotName, CommonVariables.MED_TIMEOUT);
 	}
 
 	/**
@@ -2053,7 +2039,8 @@ public class CommonFunctions {
 	// screenshots
 	public String captureScreenShot(WebDriver driver, String screenShotName) throws Exception {
 		this.logAccess.getLogger().debug("Capturing screenshot");
-		String tempScreenshotName = this.screenShotsPath + File.separatorChar + getScreenShotTime() + "_" + screenShotName + ".png";
+		String tempScreenshotName = this.screenShotsPath + File.separatorChar + getScreenShotTime() + "_"
+				+ screenShotName.replaceAll("[^-A-Za-z0-9]", "_").replace("__","_")  + ".png";
 
 		File screenshot = ((TakesScreenshot) driver).getScreenshotAs(OutputType.FILE);
 		FileUtils.copyFile(screenshot, new File(tempScreenshotName));
@@ -2091,7 +2078,8 @@ public class CommonFunctions {
 					+ getScreenShotTime() + "_" + screenShotName;
 			folderFileUtil.createFolder(tempScreenShotsFolderName);
 			capturePageChunks(driver, tempScreenShotsFolderName);
-			return mergeImagesToSingleImage(tempScreenShotsFolderName, screenShotName + ".png");
+			return mergeImagesToSingleImage(tempScreenShotsFolderName,
+					screenShotName.replaceAll("[^-A-Za-z0-9]", "_").replace("__","_")  + ".png");
 		}
 	}
 
@@ -2130,7 +2118,8 @@ public class CommonFunctions {
 			folderFileUtil.createFolder(tempScreenShotsFolderName);
 
 			capturePageChunks(driver, tempScreenShotsFolderName, headerElement, true);
-			return mergeImagesToSingleImage(tempScreenShotsFolderName, screenShotName + ".png");
+			return mergeImagesToSingleImage(tempScreenShotsFolderName,
+					screenShotName.replaceAll("[^-A-Za-z0-9]", "_").replace("__","_")  + ".png");
 		}
 
 	}
@@ -2532,6 +2521,7 @@ public class CommonFunctions {
 										int maxTimeout) throws Exception {
 		// driver.manage().timeouts().implicitlyWait(0, TimeUnit.SECONDS);
 		WebDriverWait wait = new WebDriverWait(driver, maxTimeout);
+
 		WebElement returnElement;
 		switch (expectedCondition) {
 			case CLICKABLE: